196: 멀티 진행 페이지

See merge request s10-final/S10P31A704!65

+++ wadada/lib/screens/singlerunpage/component/time_bar.dart
@@ -5,7 +5,7 @@ import 'package:percent_indicator/percent_indicator.dart';
 class TimeBar extends StatefulWidget{
   // final double dist;
   // const ProgressBar({super.key, required this.dist});
-  final double initialTime;
+  final int initialTime;
   final double elapsedTime;
 
   const TimeBar({super.key, required this.initialTime, required this.elapsedTime});
